[refactor] 將老師後台課程資訊頁的 component 切出
