+++ skilfor/src/pages/TeacherManagePage/components/PageStyle.js
@@ -1,0 +1,34 @@
+import styled from "styled-components";
+
+export const RowContainer = styled.div`
+  display: flex;
+`;
+
+export const SectionText = styled.h3`
+  font-size: 1.3rem;
+  color: ${(props) => props.theme.colors.green_dark};
+  margin: 10px 0 20px 0;
+`;
+
+export const EditContainer = styled(RowContainer)`
+  justify-content: space-between;
+  align-items: baseline;
+  margin-top: 10px;
+`;
+
+export const EditButton = styled.button`
+  background: ${(props) => props.theme.colors.green_dark};
+  border: none;
+  color: white;
+  font-size: 1rem;
+  height: fit-content;
+  padding: 7px 14px;
+  cursor: pointer;
+  :hover {
+    opacity: 0.8;
+  }
+`;
+
+export const SubmitButton = styled(EditButton)`
+  margin-left: 15px;
+`;
